monitoring/requirements.txt
streamlit
boto3
pandas
matplotlib
seaborn
plotly-express